requirements.txt
psycopg[binary]>=3.1.18
praw>=7.7.1
python-dotenv>=1.0.0
